Strategy2

+++ non_native_test.py
@@ -43,7 +43,7 @@ seq_len = tf.not_equal(x.sg_sum(axis=2), 0.).sg_int().sg_sum(axis=1)
 logit = get_logit(x, voca_size=voca_size)
 
 probability = get_predictions(logit)
-y = tf.reduce_max(x, reduction_indices=[1])
+y = tf.to_int32(tf.reduce_max(x, reduction_indices=[1]))
 
 # ctc decoding
 #decoded, _ = tf.nn.ctc_beam_search_decoder(probability.sg_transpose(perm=[1, 0, 2]), seq_len, merge_repeated=False)
